+++ cosmos/lcd.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 	"sync"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
@@ -30,7 +31,6 @@ type LCD struct {
 	accName      string
 	accPassword  string
 	minGasPrices sdktypes.DecCoins
-	gasLimit     uint64
 
 	// local state
 	acc            *auth.BaseAccount
@@ -39,7 +39,7 @@ type LCD struct {
 }
 
 // NewLCD initializes a cosmos LCD client.
-func NewLCD(endpoint string, cdc *codec.Codec, kb keys.Keybase, chainID, accName, accPassword, minGasPrices string, gasLimit uint64) (*LCD, error) {
+func NewLCD(endpoint string, cdc *codec.Codec, kb keys.Keybase, chainID, accName, accPassword, minGasPrices string) (*LCD, error) {
 	minGasPricesDecoded, err := sdktypes.ParseDecCoins(minGasPrices)
 	if err != nil {
 		return nil, err
@@ -52,7 +52,6 @@ func NewLCD(endpoint string, cdc *codec.Codec, kb keys.Keybase, chainID, accName
 		accName:      accName,
 		accPassword:  accPassword,
 		minGasPrices: minGasPricesDecoded,
-		gasLimit:     gasLimit,
 	}, nil
 }
 
@@ -208,14 +207,39 @@ func (lcd *LCD) createAndSignTx(msgs []sdk.Msg, acc *auth.BaseAccount) (authtype
 		authutils.GetTxEncoder(lcd.cdc),
 		acc.GetAccountNumber(),
 		acc.GetSequence(),
-		lcd.gasLimit,
-		flags.DefaultGasAdjustment,
+		flags.DefaultGasLimit,
+		GasAdjustment,
 		true,
 		lcd.chainID,
 		"",
 		nil,
 		lcd.minGasPrices,
 	).WithKeybase(lcd.kb)
+
+	// calculate gas
+	if txBuilder.SimulateAndExecute() {
+		gasAdjustment := strconv.FormatFloat(txBuilder.GasAdjustment(), 'f', -1, 64)
+		req := SimulateReq{
+			BaseReq: rest.NewBaseReq(
+				acc.Address.String(),
+				"",
+				lcd.chainID,
+				"",
+				gasAdjustment,
+				acc.GetAccountNumber(),
+				acc.GetSequence(),
+				nil,
+				nil,
+				true,
+			),
+			Msgs: msgs,
+		}
+		var res rest.GasEstimateResponse
+		if err := lcd.PostBare("txs/simulate", req, &res); err != nil {
+			return authtypes.StdTx{}, err
+		}
+		txBuilder = txBuilder.WithGas(res.GasEstimate)
+	}
 
 	// create StdSignMsg
 	stdSignMsg, err := txBuilder.BuildSignMsg(msgs)